Bugfix: Work Item GH-193 - Fix to empty worksheet garbage collection when using cell caching

+++ Classes/PHPExcel/CachedObjectStorage/APC.php
@@ -61,7 +61,7 @@ class PHPExcel_CachedObjectStorage_APC extends PHPExcel_CachedObjectStorage_Cach
      * @throws  PHPExcel_Exception
      */
     protected function _storeData() {
-        if ($this->_currentCellIsDirty) {
+        if ($this->_currentCellIsDirty && !empty($this->_currentObjectID)) {
             $this->_currentObject->detach();
 
             if (!apc_store($this->_cachePrefix.$this->_currentObjectID.'.cache',serialize($this->_currentObject),$this->_cacheTime)) {
